+++ js/main.js
@@ -1,4 +1,4 @@
 import introScreen from './templates/intro';
-import showScreen from './modules/showScreen';
+import showScreen from './modules/show-screen';
 
 showScreen(introScreen);
