+++ js/templates/level-first.js
@@ -1,6 +1,6 @@
-import showScreen from '../modules/showScreen';
+import showScreen from '../modules/show-screen';
 import greeting from './greeting';
-import levelThird from './levelThird';
+import levelSecond from './level-second';
 import {createElementFromTemplate} from '../utils';
 
 const screenLayout = `<header class="header">
@@ -18,16 +18,27 @@ const screenLayout = `<header class="header">
                         </div>
                       </header>
                       <div class="game">
-                        <p class="game__task">Угадай, фото или рисунок?</p>
-                        <form class="game__content  game__content--wide">
+                        <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+                        <form class="game__content">
                           <div class="game__option">
-                            <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-                            <label class="game__answer  game__answer--photo">
+                            <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+                            <label class="game__answer game__answer--photo">
                               <input name="question1" type="radio" value="photo">
                               <span>Фото</span>
                             </label>
-                            <label class="game__answer  game__answer--wide  game__answer--paint">
+                            <label class="game__answer game__answer--paint">
                               <input name="question1" type="radio" value="paint">
+                              <span>Рисунок</span>
+                            </label>
+                          </div>
+                          <div class="game__option">
+                            <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+                            <label class="game__answer  game__answer--photo">
+                              <input name="question2" type="radio" value="photo">
+                              <span>Фото</span>
+                            </label>
+                            <label class="game__answer  game__answer--paint">
+                              <input name="question2" type="radio" value="paint">
                               <span>Рисунок</span>
                             </label>
                           </div>
@@ -38,11 +49,11 @@ const screenLayout = `<header class="header">
                             <li class="stats__result stats__result--slow"></li>
                             <li class="stats__result stats__result--fast"></li>
                             <li class="stats__result stats__result--correct"></li>
-                            <li class="stats__result stats__result--wrong"></li>
                             <li class="stats__result stats__result--unknown"></li>
-                            <li class="stats__result stats__result--slow"></li>
                             <li class="stats__result stats__result--unknown"></li>
-                            <li class="stats__result stats__result--fast"></li>
+                            <li class="stats__result stats__result--unknown"></li>
+                            <li class="stats__result stats__result--unknown"></li>
+                            <li class="stats__result stats__result--unknown"></li>
                             <li class="stats__result stats__result--unknown"></li>
                           </ul>
                         </div>
@@ -50,7 +61,7 @@ const screenLayout = `<header class="header">
                       <footer class="footer">
                         <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
                         <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-                        <div class="footer__social-links">
+                         <div class="footer__social-links">
                           <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
                           <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
                           <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
@@ -60,13 +71,17 @@ const screenLayout = `<header class="header">
 
 const screenElement = createElementFromTemplate(screenLayout);
 const backButton = screenElement.querySelector(`.back`);
-const answerButtons = [...screenElement.querySelectorAll(`.game__answer`)];
-const buttonOnClickHandler = () => {
-  showScreen(levelThird);
+const radioButtons = [...screenElement.querySelectorAll(`input[type=radio]`)];
+const radioOnChangeHandler = () => {
+  const selectedRadioButtons = radioButtons.filter((radioButton) => radioButton.checked === true);
+
+  if (selectedRadioButtons.length === 2) {
+    showScreen(levelSecond);
+  }
 };
 
-answerButtons.forEach((answerButton) => {
-  answerButton.addEventListener(`click`, buttonOnClickHandler);
+radioButtons.forEach((radioButton) => {
+  radioButton.addEventListener(`change`, radioOnChangeHandler);
 });
 
 backButton.addEventListener(`click`, () => {
