+++ js/templates/level-second.js
@@ -1,6 +1,6 @@
-import showScreen from '../modules/showScreen';
+import showScreen from '../modules/show-screen';
 import greeting from './greeting';
-import statistics from './statistics';
+import levelThird from './level-third';
 import {createElementFromTemplate} from '../utils';
 
 const screenLayout = `<header class="header">
@@ -18,16 +18,18 @@ const screenLayout = `<header class="header">
                         </div>
                       </header>
                       <div class="game">
-                        <p class="game__task">Найдите рисунок среди изображений</p>
-                        <form class="game__content  game__content--triple">
+                        <p class="game__task">Угадай, фото или рисунок?</p>
+                        <form class="game__content  game__content--wide">
                           <div class="game__option">
-                            <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-                          </div>
-                          <div class="game__option  game__option--selected">
-                            <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-                          </div>
-                          <div class="game__option">
-                            <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+                            <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+                            <label class="game__answer  game__answer--photo">
+                              <input name="question1" type="radio" value="photo">
+                              <span>Фото</span>
+                            </label>
+                            <label class="game__answer  game__answer--wide  game__answer--paint">
+                              <input name="question1" type="radio" value="paint">
+                              <span>Рисунок</span>
+                            </label>
                           </div>
                         </form>
                         <div class="stats">
@@ -58,9 +60,9 @@ const screenLayout = `<header class="header">
 
 const screenElement = createElementFromTemplate(screenLayout);
 const backButton = screenElement.querySelector(`.back`);
-const answerButtons = [...screenElement.querySelectorAll(`.game__option`)];
+const answerButtons = [...screenElement.querySelectorAll(`.game__answer`)];
 const buttonOnClickHandler = () => {
-  showScreen(statistics);
+  showScreen(levelThird);
 };
 
 answerButtons.forEach((answerButton) => {
